Do not send AXSDEF command to Gx6K controllers
All GT6K controllers are stepper controllers whereas all GV6K are
servo controllers. In contrast to the 6Ks they cannot be configured
to be one or the other. Check the controller model using the TREV
command to detect if we are connected to a GT6K/GV6K controller and
do not send the AXSDEF command on these devices (it is not supported
on these and causes error messages).